user_docker/user_manage/requirements.txt
# Django核心
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Kubernetes集成
kubernetes==28.1.0
PyYAML==6.0.1


mysqlclient>=2.1.0
celery>=5.2.0,<6.0.0
redis>=4.0.0,<5.0.0
billiard>=3.6.4.0
kombu>=5.2.0,<6.0.0
# redis==5.0.1

# 认证与授权
PyJWT==2.8.0
cryptography==41.0.7

# HTTP客户端
requests==2.31.0
aiohttp==3.9.1



# 容器化
gunicorn==21.2.0

# 开发工具
python-decouple==3.8